# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Pickleball Club Ladder API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /challenges/{id}:
    get:
      operationId: getChallenge
      tags: [challenges]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Challenge
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChallengeWithTeams"
...
  /availability/{challengeId}:
    get:
      operationId: getAvailability
      tags: [availability]
      summary: Get availability for a challenge
      parameters:
        - name: challengeId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Availability
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AvailabilityResponse"
    post:
      operationId: submitAvailability
      tags: [availability]
      summary: Submit availability for a challenge
      parameters:
        - name: challengeId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitAvailabilityBody"
      responses:
        "200":
          description: Availability submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AvailabilityResponse"

  # MATCHES
...
components:
  schemas:
...
    Player:
      type: object
      properties:
        id:
          type: string
        fullName:
          type: string
        email:
          type: string
        phone:
          type: string
        role:
          type: string
          enum: [player, admin]
        isActive:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - fullName
        - email
        - role
        - isActive
        - createdAt
...
    Season:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        startDate:
          type: string
        endDate:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - name
        - startDate
        - endDate
        - isActive
        - createdAt
...
    Team:
      type: object
      properties:
        id:
          type: string
        seasonId:
          type: string
        player1Id:
          type: string
        player2Id:
          type: string
        teamName:
          type: string
        status:
          type: string
          enum: [pending, active, inactive]
        createdAt:
          type: string
      required:
        - id
        - seasonId
        - player1Id
        - player2Id
        - teamName
        - status
        - createdAt
    TeamWithPlayers:
      allOf:
        - $ref: "#/components/schemas/Team"
        - type: object
          properties:
            player1:
              $ref: "#/components/schemas/Player"
            player2:
              $ref: "#/components/schemas/Player"
            standing:
              $ref: "#/components/schemas/LadderStanding"
            season:
              $ref: "#/components/schemas/Season"
...
    LadderStanding:
      type: object
      properties:
        id:
          type: string
        seasonId:
          type: string
        teamId:
          type: string
        position:
          type: integer
        wins:
          type: integer
        losses:
          type: integer
        lastMatchDate:
          type: string
        createdAt:
          type: string
      required:
        - id
        - seasonId
        - teamId
        - position
        - wins
        - losses
        - createdAt
...
    Challenge:
      type: object
      properties:
        id:
          type: string
        seasonId:
          type: string
        challengerTeamId:
          type: string
        challengedTeamId:
          type: string
        status:
          type: string
          enum: [pending, accepted, scheduling, scheduled, completed, cancelled, disputed]
        createdAt:
          type: string
      required:
        - id
        - seasonId
        - challengerTeamId
        - challengedTeamId
        - status
        - createdAt
    ChallengeWithTeams:
      allOf:
        - $ref: "#/components/schemas/Challenge"
        - type: object
          properties:
            challengerTeam:
              $ref: "#/components/schemas/TeamWithPlayers"
            challengedTeam:
              $ref: "#/components/schemas/TeamWithPlayers"
            match:
              $ref: "#/components/schemas/MatchWithDetails"
            myTeamId:
              type: string
            overlappingSlots:
              type: array
              items:
                $ref: "#/components/schemas/TimeSlot"
            challengerAvailabilitySubmitted:
              type: boolean
            challengedAvailabilitySubmitted:
              type: boolean
...
    TimeSlot:
      type: object
      properties:
        date:
          type: string
        times:
          type: array
          items:
            type: string
      required:
        - date
        - times

    # Availability
    AvailabilitySubmission:
      type: object
      properties:
        id:
          type: string
        challengeId:
          type: string
        teamId:
          type: string
        slots:
          type: array
          items:
            $ref: "#/components/schemas/TimeSlot"
        submittedAt:
          type: string
      required:
        - id
        - challengeId
        - teamId
        - slots
        - submittedAt
    SubmitAvailabilityBody:
      type: object
      properties:
        slots:
          type: array
          items:
            $ref: "#/components/schemas/TimeSlot"
      required:
        - slots
    AvailabilityResponse:
      type: object
      properties:
        challengerAvailability:
          $ref: "#/components/schemas/AvailabilitySubmission"
        challengedAvailability:
          $ref: "#/components/schemas/AvailabilitySubmission"
        overlappingSlots:
          type: array
          items:
            $ref: "#/components/schemas/TimeSlot"
        hasOverlap:
          type: boolean
      required:
        - hasOverlap

    # Match
    Match:
      type: object
      properties:
        id:
          type: string
        challengeId:
          type: string
        scheduledDate:
          type: string
        scheduledTime:
          type: string
        courtLocation:
          type: string
        bookedByTeamId:
          type: string
        status:
          type: string
          enum: [scheduled, completed, disputed]
        createdAt:
          type: string
      required:
        - id
        - challengeId
        - scheduledDate
        - scheduledTime
        - courtLocation
        - status
        - createdAt
    MatchScore:
      type: object
      properties:
        id:
          type: string
        matchId:
          type: string
        gameNumber:
          type: integer
        team1Score:
          type: integer
        team2Score:
          type: integer
      required:
        - id
        - matchId
        - gameNumber
        - team1Score
        - team2Score
    MatchResult:
      type: object
      properties:
        id:
          type: string
        matchId:
          type: string
        winnerTeamId:
          type: string
        loserTeamId:
          type: string
        submittedByTeamId:
          type: string
        confirmedByTeamId:
          type: string
        submittedAt:
          type: string
        confirmedAt:
          type: string
        autoConfirmed:
          type: boolean
        disputeReason:
          type: string
        disputeResolved:
          type: boolean
      required:
        - id
        - matchId
        - winnerTeamId
        - loserTeamId
        - submittedByTeamId
        - submittedAt
        - autoConfirmed
        - disputeResolved
    MatchWithDetails:
      allOf:
        - $ref: "#/components/schemas/Match"
        - type: object
          properties:
            challenge:
              $ref: "#/components/schemas/ChallengeWithTeams"
            scores:
              type: array
              items:
                $ref: "#/components/schemas/MatchScore"
            result:
              $ref: "#/components/schemas/MatchResult"
...